Move service_container.py to src/core/ for systematic project structure - Created src/core/ directory for architectural components - Moved service_container.py from root to src/core/service_container.py - Updated all import statements across codebase - Updated Dockerfile to copy entire src/ directory - Root directory now contains only essential application files - All tests pass with new structure (25/25 ✅)
