+++ tests/unit/test_main.py
@@ -20,7 +20,7 @@ sys.modules["telegram"] = Mock()
 sys.modules["telegram.ext"] = Mock()
 
 from main import create_app  # noqa: E402
-from service_container import reset_service_container  # noqa: E402
+from src.core.service_container import reset_service_container  # noqa: E402
 
 
 @pytest.fixture
